+++ gax/test/browser-test/karma.conf.js
@@ -22,7 +22,7 @@ const isDocker = require('is-docker')();
 const webpackConfig = require('./webpack.config.js');
 webpackConfig.performance = {hints: false};
 delete webpackConfig.entry;
-// eslint-disable-next-line node/no-missing-require
+// eslint-disable-next-line n/no-missing-require
 process.env.CHROME_BIN = require('puppeteer').executablePath();
 
 module.exports = function (config) {
@@ -38,7 +38,7 @@ module.exports = function (config) {
       childProcess: {
         path: path.relative(
           process.cwd(),
-          // eslint-disable-next-line node/no-missing-require
+          // eslint-disable-next-line n/no-missing-require
           require.resolve('showcase-server/build/src/index.js')
         ),
         args: [],