+++ gax/test/unit/streaming.ts
@@ -21,7 +21,12 @@ import * as sinon from 'sinon';
 import {afterEach, describe, it} from 'mocha';
 import {PassThrough} from 'stream';
 
-import {GaxCallStream, GRPCCall, RequestType} from '../../src/apitypes';
+import {
+  GaxCallStream,
+  GRPCCall,
+  RequestType,
+  CancellableStream,
+} from '../../src/apitypes';
 import {createApiCall} from '../../src/createApiCall';
 import {StreamingApiCaller} from '../../src/streamingCalls/streamingApiCaller';
 import * as gax from '../../src/gax';
@@ -234,8 +239,7 @@ describe('streaming', () => {
   });
 
   it('cancels in the middle', done => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    function schedulePush(s: any, c: number) {
+    function schedulePush(s: CancellableStream, c: number) {
       const intervalId = setInterval(() => {
         s.push(c);
         c++;
@@ -248,10 +252,9 @@ describe('streaming', () => {
     function func() {
       const s = new PassThrough({
         objectMode: true,
-      });
+      }) as unknown as CancellableStream;
       schedulePush(s, 0);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (s as any).cancel = () => {
+      s.cancel = () => {
         s.end();
         s.emit('error', cancelError);
       };
@@ -264,6 +267,70 @@ describe('streaming', () => {
       //@ts-ignore
       func,
       streaming.StreamType.SERVER_STREAMING
+    );
+    const s = apiCall(
+      {},
+      {
+        retry: gax.createRetryOptions([5], {
+          initialRetryDelayMillis: 100,
+          retryDelayMultiplier: 1.2,
+          maxRetryDelayMillis: 1000,
+          rpcTimeoutMultiplier: 1.5,
+          maxRpcTimeoutMillis: 3000,
+          maxRetries: 0,
+        }),
+      }
+    );
+    let counter = 0;
+    const expectedCount = 5;
+    s.on('data', data => {
+      assert.strictEqual(data, counter);
+      counter++;
+      if (counter === expectedCount) {
+        s.cancel();
+      } else if (counter > expectedCount) {
+        done(new Error('should not reach'));
+      }
+    });
+    s.on('error', err => {
+      assert.strictEqual(err, cancelError);
+      done();
+    });
+  });
+  it('cancels in the middle when new retries are enabled', done => {
+    function schedulePush(s: CancellableStream, c: number) {
+      const intervalId = setInterval(() => {
+        s.push(c);
+        c++;
+      }, 10);
+      s.on('finish', () => {
+        clearInterval(intervalId);
+      });
+    }
+    const cancelError = new Error('cancelled');
+    function func() {
+      const s = new PassThrough({
+        objectMode: true,
+      }) as unknown as CancellableStream;
+      schedulePush(s, 0);
+      s.cancel = () => {
+        s.end();
+        s.emit('error', cancelError);
+      };
+      setImmediate(() => {
+        s.emit('metadata');
+      });
+      setImmediate(() => {
+        s.emit('status');
+      });
+      return s;
+    }
+    const apiCall = createApiCallStreaming(
+      //@ts-ignore
+      func,
+      streaming.StreamType.SERVER_STREAMING,
+      false,
+      true // gax native retries
     );
     const s = apiCall(
       {},
@@ -309,7 +376,9 @@ describe('streaming', () => {
       const s = new PassThrough({
         objectMode: true,
       });
-      s.push(null);
+      setImmediate(() => {
+        s.push(null);
+      });
       setImmediate(() => {
         s.emit('metadata', responseMetadata);
       });
@@ -829,11 +898,91 @@ describe('streaming', () => {
       done();
     });
   });
-  it('emit transient error on second or later error when new retries are enabled', done => {
-    // stubbing cancel is needed because PassThrough doesn't have
-    // a cancel method and cancel is called as part of the retry
-    const cancelStub = sinon.stub(streaming.StreamProxy.prototype, 'cancel');
+  it('emit transient error message if neither maxRetries nor totaltimeout are defined when new retries are enabled', done => {
+    const errorInfoObj = {
+      reason: 'SERVICE_DISABLED',
+      domain: 'googleapis.com',
+      metadata: {
+        consumer: 'projects/455411330361',
+        service: 'translate.googleapis.com',
+      },
+    };
+    const errorProtoJson = require('../../protos/status.json');
+    const root = protobuf.Root.fromJSON(errorProtoJson);
+    const errorInfoType = root.lookupType('ErrorInfo');
+    const buffer = errorInfoType.encode(errorInfoObj).finish() as Buffer;
+    const any = {
+      type_url: 'type.googleapis.com/google.rpc.ErrorInfo',
+      value: buffer,
+    };
+    const status = {code: 3, message: 'test', details: [any]};
+    const Status = root.lookupType('google.rpc.Status');
+    const status_buffer = Status.encode(status).finish() as Buffer;
+    const metadata = new Metadata();
+    metadata.set('grpc-status-details-bin', status_buffer);
+    const error = Object.assign(new GoogleError('test error'), {
+      code: 3,
+      details: 'Failed to read',
+      metadata: metadata,
+    });
 
+    const spy = sinon.spy((...args: Array<{}>) => {
+      assert.strictEqual(args.length, 3);
+      const s = new PassThrough({
+        objectMode: true,
+      });
+      s.push(null);
+      setImmediate(() => {
+        // emits an error not in our included retry codes
+        s.emit('error', error);
+      });
+      setImmediate(() => {
+        s.emit('status', status);
+      });
+
+      return s;
+    });
+    const apiCall = createApiCallStreaming(
+      spy,
+      streaming.StreamType.SERVER_STREAMING,
+      false,
+      true // new retry behavior enabled
+    );
+
+    const s = apiCall(
+      {},
+      {
+        retry: gax.createRetryOptions([5], {
+          initialRetryDelayMillis: 100,
+          retryDelayMultiplier: 1.2,
+          maxRetryDelayMillis: 1000,
+          rpcTimeoutMultiplier: 1.5,
+          maxRpcTimeoutMillis: 3000,
+          // neither maxRetries nor totalTimeoutMillis is defined
+        }),
+      }
+    );
+
+    s.on('error', err => {
+      s.pause();
+      s.destroy();
+
+      assert(err instanceof GoogleError);
+      assert.deepStrictEqual(err.message, 'test error');
+      assert.deepStrictEqual(
+        err.note,
+        'Exception occurred in retry method that was not classified as transient'
+      );
+      assert.strictEqual(err.domain, errorInfoObj.domain);
+      assert.strictEqual(err.reason, errorInfoObj.reason);
+      assert.strictEqual(
+        JSON.stringify(err.errorInfoMetadata),
+        JSON.stringify(errorInfoObj.metadata)
+      );
+      done();
+    });
+  });
+  it('emit transient error on second or later error when new retries are enabled', done => {
     const errorInfoObj = {
       reason: 'SERVICE_DISABLED',
       domain: 'googleapis.com',
@@ -877,7 +1026,9 @@ describe('streaming', () => {
         case 0:
           e = error;
 
-          s.push(null);
+          setImmediate(() => {
+            s.push('hello');
+          });
           setImmediate(() => {
             s.emit('error', e); // is included in our retry codes
           });
@@ -890,7 +1041,9 @@ describe('streaming', () => {
         case 1:
           e = error2; // is not in our retry codes
 
-          s.push(null);
+          setImmediate(() => {
+            s.push('world');
+          });
           setImmediate(() => {
             s.emit('error', e);
           });
@@ -902,6 +1055,10 @@ describe('streaming', () => {
 
           return s;
         default:
+          // should not reach this
+          setImmediate(() => {
+            s.emit('status', status);
+          });
           setImmediate(() => {
             s.emit('end');
           });
@@ -943,11 +1100,9 @@ describe('streaming', () => {
         JSON.stringify(err.errorInfoMetadata),
         JSON.stringify(errorInfoObj.metadata)
       );
-      assert.strictEqual(cancelStub.callCount, 1);
       done();
     });
   });
-
   it('emit error and retry once', done => {
     const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
       code: 14,
@@ -963,10 +1118,10 @@ describe('streaming', () => {
         objectMode: true,
       });
       setImmediate(() => {
-        s.push('Hello');
-        s.push('World');
         switch (counter) {
           case 0:
+            s.push('Hello');
+            s.push('World');
             s.emit('error', firstError);
             counter++;
             break;
@@ -1013,12 +1168,7 @@ describe('streaming', () => {
     });
   });
 
-  it('emit error and retry twice with shouldRetryFn', done => {
-    // stubbing cancel is needed because PassThrough doesn't have
-    // a cancel method and cancel is called as part of the retry
-    sinon.stub(streaming.StreamProxy.prototype, 'cancel').callsFake(() => {
-      done();
-    });
+  it('emit error, retry twice, and succeed with shouldRetryFn', done => {
     const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
       code: 14,
       details: 'UNAVAILABLE',
@@ -1034,6 +1184,8 @@ describe('streaming', () => {
       switch (counter) {
         case 0:
           setImmediate(() => {
+            s.push('Hello');
+            s.push('World');
             s.emit('error', firstError);
           });
           setImmediate(() => {
@@ -1043,6 +1195,8 @@ describe('streaming', () => {
           return s;
         case 1:
           setImmediate(() => {
+            s.push('testing');
+            s.push('retries');
             s.emit('error', firstError);
           });
           setImmediate(() => {
@@ -1051,6 +1205,14 @@ describe('streaming', () => {
           counter++;
           return s;
         default:
+          setImmediate(() => {
+            s.emit('metadata');
+          });
+          // grpc streams always emit status
+          setImmediate(() => {
+            s.emit('status');
+          });
+          // emit end because there is no more data
           setImmediate(() => {
             s.emit('end');
           });
@@ -1085,16 +1247,18 @@ describe('streaming', () => {
         ),
       }
     );
-
+    const finalData: string[] = [];
+    s.on('data', data => {
+      finalData.push(data);
+    });
     s.on('end', () => {
       s.destroy();
       assert.strictEqual(counter, 2);
+      assert.strictEqual(finalData.join(' '), 'Hello World testing retries');
+      done();
     });
   });
   it('retries using resumption request function ', done => {
-    // stubbing cancel is needed because PassThrough doesn't have
-    // a cancel method and cancel is called as part of the retry
-    sinon.stub(streaming.StreamProxy.prototype, 'cancel');
     const receivedData: string[] = [];
     const error = Object.assign(new GoogleError('test error'), {
       code: 14,
@@ -1111,8 +1275,11 @@ describe('streaming', () => {
       });
       switch (arg) {
         case 0:
-          s.push('Hello');
-          s.push('World');
+          setImmediate(() => {
+            s.push('Hello');
+            s.push('World');
+          });
+
           setImmediate(() => {
             s.emit('metadata');
           });
@@ -1134,10 +1301,16 @@ describe('streaming', () => {
           });
           return s;
         case 2:
-          s.push('testing');
-          s.push('retries');
+          setImmediate(() => {
+            s.push('testing');
+            s.push('retries');
+          });
+
           setImmediate(() => {
             s.emit('metadata');
+          });
+          setImmediate(() => {
+            s.emit('status');
           });
           setImmediate(() => {
             s.emit('end');
@@ -1157,10 +1330,10 @@ describe('streaming', () => {
       true // new retry behavior enabled
     );
     // resumption strategy is to pass a different arg to the function
-    const getResumptionRequestFn = (originalRequest: RequestType) => {
+    const getResumptionRequestFn = sinon.spy((originalRequest: RequestType) => {
       assert.strictEqual(originalRequest.arg, 0);
       return {arg: 2};
-    };
+    });
     const s = apiCall(
       {arg: 0},
       {
@@ -1194,13 +1367,11 @@ describe('streaming', () => {
         receivedData.join(' '),
         'Hello World testing retries'
       );
+      assert.strictEqual(getResumptionRequestFn.callCount, 1);
       done();
     });
   });
   it('errors when there is a resumption request function an gaxStreamingRetries is not enabled', done => {
-    // stubbing cancel is needed because PassThrough doesn't have
-    // a cancel method and cancel is called as part of the retry
-    sinon.stub(streaming.StreamProxy.prototype, 'cancel');
     const error = Object.assign(new GoogleError('test error'), {
       code: 14,
       details: 'UNAVAILABLE',
@@ -1301,9 +1472,11 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
   afterEach(() => {
     sinon.restore();
   });
-
-  it('server streaming call retries until exceeding max retries', done => {
-    const retrySpy = sinon.spy(streaming.StreamProxy.prototype, 'retry');
+  it('server streaming call retries until exceeding timeout and surfaces underlying error', done => {
+    const retrySpy = sinon.spy(
+      streaming.StreamProxy.prototype,
+      'throwIfMaxRetriesOrTotalTimeoutExceeded'
+    );
     const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
       code: 14,
       details: 'UNAVAILABLE',
@@ -1315,7 +1488,70 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
       const s = new PassThrough({
         objectMode: true,
       });
-      s.push('hello');
+      setImmediate(() => {
+        s.emit('metadata');
+      });
+      setImmediate(() => {
+        s.emit('error', firstError);
+      });
+      return s;
+    });
+
+    const apiCall = createApiCallStreaming(
+      spy,
+      streaming.StreamType.SERVER_STREAMING,
+      false,
+      true
+    );
+
+    const call = apiCall(
+      {},
+      {
+        retry: gax.createRetryOptions([14], {
+          initialRetryDelayMillis: 100,
+          retryDelayMultiplier: 1.2,
+          maxRetryDelayMillis: 1000,
+          rpcTimeoutMultiplier: 1.5,
+          maxRpcTimeoutMillis: 3000,
+          totalTimeoutMillis: 200, // timeout that ensures it should retry at least once
+        }),
+      }
+    );
+
+    call.on('error', err => {
+      try {
+        assert(err instanceof GoogleError);
+        if (err.code !== 14) {
+          // ignore the error we are expecting
+          assert.strictEqual(err.code, 4);
+          assert.notStrictEqual(retrySpy.callCount, 0); // it MUST retry at least once
+          assert.strictEqual(
+            err.message,
+            'Total timeout of API exceeded 200 milliseconds retrying error Error: UNAVAILABLE  before any response was received.'
+          );
+          done();
+        }
+      } catch (error: unknown) {
+        done(error);
+      }
+    });
+  });
+  it('server streaming call retries until exceeding max retries and surfaces underlying error in note', done => {
+    const retrySpy = sinon.spy(
+      streaming.StreamProxy.prototype,
+      'throwIfMaxRetriesOrTotalTimeoutExceeded'
+    );
+    const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
+      code: 14,
+      details: 'UNAVAILABLE',
+      metadata: new Metadata(),
+    });
+
+    const spy = sinon.spy((...args: Array<{}>) => {
+      assert.strictEqual(args.length, 3);
+      const s = new PassThrough({
+        objectMode: true,
+      });
       setImmediate(() => {
         s.emit('metadata');
       });
@@ -1347,24 +1583,29 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
     );
 
     call.on('error', err => {
-      assert(err instanceof GoogleError);
-      if (err.code !== 14) {
-        // ignore the error we are expecting
-        assert.strictEqual(err.code, 4);
-        // even though max retries is 2
-        // the retry function will always be called maxRetries+1
-        // the final call is where the failure happens
-        assert.strictEqual(retrySpy.callCount, 3);
-        assert.strictEqual(
-          err.message,
-          'Exceeded maximum number of retries before any response was received'
-        );
-        done();
+      try {
+        assert(err instanceof GoogleError);
+        if (err.code !== 14) {
+          // ignore the error we are expecting
+          assert.strictEqual(err.code, 4);
+          assert.strictEqual(retrySpy.callCount, 3); // we pass the first two times
+          assert.strictEqual(
+            err.message,
+            'Exceeded maximum number of retries retrying error Error: UNAVAILABLE before any response was received'
+          );
+          done();
+        }
+      } catch (error: unknown) {
+        done(error);
       }
     });
   });
   it('does not retry when there is no shouldRetryFn and retryCodes is an empty array', done => {
-    const retrySpy = sinon.spy(streaming.StreamProxy.prototype, 'retry');
+    // we don't call the timeout/max retry check on non retryable error codes
+    const retrySpy = sinon.spy(
+      streaming.StreamProxy.prototype,
+      'throwIfMaxRetriesOrTotalTimeoutExceeded'
+    );
     const firstError = Object.assign(new GoogleError('UNAVAILABLE'), {
       code: 14,
       details: 'UNAVAILABLE',
@@ -1468,7 +1709,7 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
         assert.strictEqual(err.code, 4);
         assert.strictEqual(
           err.message,
-          'Total timeout of API exceeded 10 milliseconds before any response was received.'
+          'Total timeout of API exceeded 10 milliseconds retrying error Error: UNAVAILABLE  before any response was received.'
         );
         done();
       }
@@ -1476,7 +1717,7 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
   });
   it('allows custom CallOptions.retry settings with shouldRetryFn instead of retryCodes and new retry behavior', done => {
     sinon
-      .stub(streaming.StreamProxy.prototype, 'forwardEventsWithRetries')
+      .stub(streaming.StreamProxy.prototype, 'eventForwardHelper')
       .callsFake((stream): undefined => {
         assert(stream instanceof internal.Stream);
         done();
@@ -1496,6 +1737,7 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
       true //gaxStreamingRetries
     );
 
+    // anonymous function is a shouldRetryFn
     apiCall(
       {},
       {
@@ -1518,7 +1760,7 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
   });
   it('allows custom CallOptions.retry settings with retryCodes and new retry behavior', done => {
     sinon
-      .stub(streaming.StreamProxy.prototype, 'forwardEventsWithRetries')
+      .stub(streaming.StreamProxy.prototype, 'eventForwardHelper')
       .callsFake((stream): undefined => {
         assert(stream instanceof internal.Stream);
         done();
@@ -1554,11 +1796,19 @@ describe('handles server streaming retries in gax when gaxStreamingRetries is en
   });
   it('allows the user to pass a custom resumption strategy', done => {
     sinon
-      .stub(streaming.StreamProxy.prototype, 'forwardEventsWithRetries')
-      .callsFake((stream, retry): undefined => {
-        assert(stream instanceof internal.Stream);
-        assert(retry.getResumptionRequestFn instanceof Function);
+      // typecasting to any is a workaround for stubbing private functions in sinon
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      .stub(streaming.StreamProxy.prototype, 'newStreamingRetryRequest' as any)
+      .callsFake((opts): CancellableStream => {
+        // @ts-ignore errors on unknown type because newStreamingRetryRequest is a private function
+        assert(opts.retry.getResumptionRequestFn instanceof Function);
         done();
+        const returnStream = new PassThrough() as unknown as CancellableStream;
+        returnStream.cancel = () => {
+          returnStream.destroy();
+        };
+        // we have to return something like newStreamingRetryRequest does
+        return returnStream;
       });
     const spy = sinon.spy((...args: Array<{}>) => {
       assert.strictEqual(args.length, 3);